cv_chat_api/pyproject.toml
[project]
name = "cv_chat_api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "openai>=2.15.0",
    "perplexityai>=0.23.0",
    "pypdf>=6.6.0",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
